feat: proper error handling & error propagation
